Separate QueryProperty into QueryableProperty and QueryBuilderProperty
