+++ src/QueryDocuments/QueryBuilderPropertyData.ts
@@ -1,0 +1,7 @@
+import { QueryablePropertyData } from "./QueryablePropertyData";
+import { QueryBuilderProperty } from "./QueryBuilderProperty";
+
+
+export interface QueryBuilderPropertyData extends QueryablePropertyData {
+	parent?:QueryBuilderProperty;
+}
