+++ src/QueryDocuments/QueryContainerType.ts
@@ -1,0 +1,5 @@
+export enum QueryContainerType {
+	DOCUMENT,
+	CHILDREN,
+	MEMBERS,
+}
