--- src/QueryDocuments/QueryRootContainerType.ts
@@ -1,4 +1,0 @@
-export enum QueryRootContainerType {
-	CHILDREN,
-	MEMBERS,
-}
