+++ test/helpers/mocks/querying/MockQueryContainer.ts
@@ -2,20 +2,21 @@ import { AbstractContext } from "../../../../src/Context/AbstractContext";
 
 import { DigestedObjectSchemaProperty } from "../../../../src/ObjectSchema/DigestedObjectSchemaProperty";
 
+import { QueryBuilderProperty } from "../../../../src/QueryDocuments/QueryBuilderProperty";
 import { QueryContainer } from "../../../../src/QueryDocuments/QueryContainer";
-import { QueryProperty2 } from "../../../../src/QueryDocuments/QueryProperty2";
 
 
 export class MockQueryContainer extends QueryContainer {
-	readonly _queryProperty:QueryProperty2;
+	readonly _queryProperty:QueryBuilderProperty;
 
 	constructor( context:AbstractContext<any, any, any> ) {
 		super( context );
 
-		this._queryProperty = new QueryProperty2( {
+		this._queryProperty = new QueryBuilderProperty( {
 			queryContainer: this,
 			name: "property",
 			definition: new DigestedObjectSchemaProperty(),
+			optional: true,
 		} );
 	}
 }
